reworked processor-session-procedure
sessions are now initiated outside the processors run-method. That means
their __init__ is executed in a non-multiprocessing-szenario and are
safe to do any db-stuff.
Now the only part that are executed by fabric is the process-method.

We also pass the request to Session.__init__ and .rework. That allows us
to use the message-engine there.