Overworked session-interruption
Introduced the concept of a soft-interruption, which waits for the execute-method to be done. The background is, that a shell-proc could not be interrupted if no pty is in use. In this case it makes most sense to just wait for the current shell-proc. The execute-method will be protected from interruption so we can fetch the response of fabric's run-method at least.